requirements.txt
streamlit
faiss-cpu
openai
numpy
pickle5
